Copy maps using Go 1.21's new "maps" package
